+++ Source/Scene/Scene.js
@@ -393,6 +393,7 @@ define([
     };
 
     var scratchOccluderBoundingSphere = new BoundingSphere();
+    var scratchOccluder;
 
     function getOccluder(scene) {
         // TODO: The occluder is the top-level central body. When we add
@@ -401,7 +402,8 @@ define([
         if (scene.mode === SceneMode.SCENE3D && defined(cb)) {
             var ellipsoid = cb.getEllipsoid();
             scratchOccluderBoundingSphere.radius = ellipsoid.getMinimumRadius();
-            return new Occluder(scratchOccluderBoundingSphere, scene._camera.positionWC);
+            scratchOccluder = Occluder.fromBoundingSphere(scratchOccluderBoundingSphere, scene._camera.positionWC, scratchOccluder);
+            return scratchOccluder;
         }
 
         return undefined;
@@ -894,6 +896,9 @@ define([
     };
 
     var orthoPickingFrustum = new OrthographicFrustum();
+    var scratchBufferDimensions = new Cartesian2();
+    var scratchPixelSize = new Cartesian2();
+
     function getPickOrthographicCullingVolume(scene, drawingBufferPosition, width, height) {
         var context = scene._context;
         var camera = scene._camera;
@@ -912,7 +917,10 @@ define([
         position.y += x;
         position.z += y;
 
-        var pixelSize = frustum.getPixelSize(new Cartesian2(drawingBufferWidth, drawingBufferHeight));
+        scratchBufferDimensions.x = drawingBufferWidth;
+        scratchBufferDimensions.y = drawingBufferHeight;
+
+        var pixelSize = frustum.getPixelSize(scratchBufferDimensions, undefined, scratchPixelSize);
 
         var ortho = orthoPickingFrustum;
         ortho.right = pixelSize.x * 0.5;
@@ -926,6 +934,7 @@ define([
     }
 
     var perspPickingFrustum = new PerspectiveOffCenterFrustum();
+
     function getPickPerspectiveCullingVolume(scene, drawingBufferPosition, width, height) {
         var context = scene._context;
         var camera = scene._camera;
@@ -944,7 +953,10 @@ define([
         var xDir = x * near * tanTheta;
         var yDir = y * near * tanPhi;
 
-        var pixelSize = frustum.getPixelSize(new Cartesian2(drawingBufferWidth, drawingBufferHeight));
+        scratchBufferDimensions.x = drawingBufferWidth;
+        scratchBufferDimensions.y = drawingBufferHeight;
+
+        var pixelSize = frustum.getPixelSize(scratchBufferDimensions, undefined, scratchPixelSize);
         var pickWidth = pixelSize.x * width * 0.5;
         var pickHeight = pixelSize.y * height * 0.5;
 
@@ -972,6 +984,7 @@ define([
     var rectangleHeight = 3.0;
     var scratchRectangle = new BoundingRectangle(0.0, 0.0, rectangleWidth, rectangleHeight);
     var scratchColorZero = new Color(0.0, 0.0, 0.0, 0.0);
+    var scratchPosition = new Cartesian2();
 
     /**
      * Returns an object with a `primitive` property that contains the first (top) primitive in the scene
@@ -996,7 +1009,7 @@ define([
         var us = this.getUniformState();
         var frameState = this._frameState;
 
-        var drawingBufferPosition = SceneTransforms.transformWindowToDrawingBuffer(context, windowPosition);
+        var drawingBufferPosition = SceneTransforms.transformWindowToDrawingBuffer(context, windowPosition, scratchPosition);
 
         if (!defined(this._pickFramebuffer)) {
             this._pickFramebuffer = context.createPickFramebuffer();